Update agent's Eventbus to synchronously dispatch Initialize and Destroy Events
Correction of issue #24 and sarl/sarl#56.
Integration of the AsyncSyncEventBus to combine asynchronous dispatching
of common events and synchronous dispatching of Initialize and Destroy
Events.
Just to be sure that an agent is effectively initialized before
processing other events.
Respectively, being sure that an agent has effectively free its
ressources before uninstalling built-in skills.