Fixing ZeroMQ re-connection.
The code that permits to re-connect the ZeroMQ sockets is not put at
the correct location in the code: the code is inside a then-block that
is executed only at the first connection of a space; and this is
not the case during the re-connection.
This patch is fixing the issue #57.
The patch is tested with 2, 3, and 4 kernels on the same host (issue
#48).